+++ lib/src/base/banner.dart
@@ -3,39 +3,6 @@ import 'package:arna/arna.dart';
 /// Banner types.
 enum BannerType { information, warning, error, success, colored }
 
-/// Specify how an [ArnaBanner] was closed.
-///
-/// The [ArnaScaffoldMessengerState.showBanner] function returns a
-/// [ArnaScaffoldFeatureController]. The value of the controller's closed
-/// property is a Future that resolves to an ArnaBannerClosedReason.
-/// Applications that need to know how a banner was closed can use this value.
-///
-/// Example:
-///
-/// ```dart
-/// ArnaScaffoldMessenger.of(context).showBanner(
-///   ArnaBanner( ... )
-/// ).closed.then((ArnaBannerClosedReason reason) {
-///    ...
-/// });
-/// ```
-enum ArnaBannerClosedReason {
-  /// The banner was closed through a [SemanticsAction.dismiss].
-  dismiss,
-
-  /// The banner was closed by a user's swipe.
-  swipe,
-
-  /// The banner was closed by the [ArnaScaffoldFeatureController] close
-  /// callback or by calling [ArnaScaffoldMessengerState.hideCurrentBanner]
-  /// directly.
-  hide,
-
-  /// The banner was closed by a call to
-  /// [ArnaScaffoldMessengerState.removeCurrentBanner].
-  remove,
-}
-
 /// An Arna-styled banner.
 ///
 /// A banner displays an important, succinct message, and provides actions for
@@ -49,13 +16,16 @@ class ArnaBanner extends StatefulWidget {
   /// Creates a banner in the Arna style.
   const ArnaBanner({
     Key? key,
+    required this.showBanner,
     required this.title,
     this.subtitle,
     required this.actions,
-    this.animation,
     this.accentColor,
     this.bannerType = BannerType.information,
   }) : super(key: key);
+
+  /// Whether to show banner or not.
+  final bool showBanner;
 
   /// The primary content of the banner.
   final String title;
@@ -66,57 +36,58 @@ class ArnaBanner extends StatefulWidget {
   /// The set of actions that are displayed at the trailing side of the banner.
   final List<Widget>? actions;
 
-  /// The animation driving the entrance and exit of the banner when presented
-  /// by the [ArnaScaffoldMessenger].
-  final Animation<double>? animation;
-
   /// The indicator color of the banner.
   final Color? accentColor;
 
   /// The type of the banner.
   final BannerType bannerType;
 
-  // API for ArnaScaffoldMessengerState.showBanner():
-
-  /// Creates an animation controller useful for driving a banner's entrance
-  /// and exit animation.
-  static AnimationController createAnimationController({
-    required TickerProvider vsync,
-  }) {
-    return AnimationController(
-      duration: Styles.basicDuration,
-      debugLabel: 'ArnaBanner',
-      vsync: vsync,
-    );
-  }
-
-  /// Creates a copy of this banner but with the animation replaced with the
-  /// given animation.
-  ///
-  /// If the original banner lacks a key, the newly created banner will use the
-  /// given fallback key.
-  ArnaBanner withAnimation(Animation<double> newAnimation, {Key? fallbackKey}) {
-    return ArnaBanner(
-      key: key ?? fallbackKey,
-      title: title,
-      subtitle: subtitle,
-      actions: actions,
-      animation: newAnimation,
-      accentColor: accentColor,
-      bannerType: bannerType,
-    );
-  }
-
   @override
   _ArnaBannerState createState() => _ArnaBannerState();
 }
 
 /// The [State] for a [ArnaBanner].
-class _ArnaBannerState extends State<ArnaBanner> {
+class _ArnaBannerState extends State<ArnaBanner>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _controller;
+  late Animation<double> _animation;
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = AnimationController(
+      duration: Styles.basicDuration,
+      debugLabel: 'ArnaBanner',
+      vsync: this,
+    );
+    _animation = CurvedAnimation(parent: _controller, curve: Styles.basicCurve);
+    if (widget.showBanner) _controller.forward();
+  }
+
+  @override
+  void didUpdateWidget(ArnaBanner oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    if (widget.showBanner != oldWidget.showBanner) {
+      switch (_controller.status) {
+        case AnimationStatus.completed:
+        case AnimationStatus.dismissed:
+          widget.showBanner ? _controller.forward() : _controller.reverse();
+          break;
+        case AnimationStatus.forward:
+        case AnimationStatus.reverse:
+          break;
+      }
+    }
+  }
+
+  @override
+  void dispose() {
+    _controller.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
-    assert(debugCheckHasMediaQuery(context));
-
     IconData icon;
     Color accent;
     switch (widget.bannerType) {
@@ -150,9 +121,6 @@ class _ArnaBannerState extends State<ArnaBanner> {
     Widget banner = Semantics(
       container: true,
       liveRegion: true,
-      onDismiss: () => ArnaScaffoldMessenger.of(context).removeCurrentBanner(
-        reason: ArnaBannerClosedReason.dismiss,
-      ),
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: <Widget>[
@@ -182,10 +150,7 @@ class _ArnaBannerState extends State<ArnaBanner> {
           ? banner
           : SizeTransition(
               axisAlignment: 1,
-              sizeFactor: CurvedAnimation(
-                parent: widget.animation!,
-                curve: Styles.basicCurve,
-              ),
+              sizeFactor: _animation,
               child: banner,
             ),
     );